organize student folder
