+++ frontend/src/components/student-experience/questions/question-elements/Image.tsx
