+++ frontend/src/components/student-experience/questions/question-elements/QuestionText.tsx
